:sparkles: add `tailwindcss` and move dirs into `src/`
