+++ src/components/styles/globals.css
