+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
-import styles from '../styles/Home.module.css'
+import styles from '../components/styles/Home.module.css'
 
 const Home: NextPage = () => {
   return (
